inspect tile script
